ProfsPortail is a collaborative platform designed for institutes to facilitate seamless communication and management among professors, Head of Departments (HOD), and students.
ProfsPortail is a Platform for the different branches of an Institute where the Professors and Head of Department can work collaborately and the students who want to know about their department and their respective Professors  they can come to this platform and see about them. In this website Student can also ask query to their respective professors. the major thing which makes this platform different from other is that Here Professors can send their leave application with only one click. Professors have only to give reason of leave and the the no. of leave days. Ai can generate their leave application and sent to The respective receivers. In this Platform Head of Department(HOD) can manage HOD and The listed subjects. One major speciality about this platform is that Professors can chat with HOD directly through Chat window. also HOD can see the reports of the Professors and their department between the days or months HOD want.

Other functionalities are AI based daily routine scheduler and Professor's Community Window.

Co-Authored-By: Dhanya Dwivedi <134362631+dhanya-17@users.noreply.github.com>